+++ assemblerflow/generator/components/assembly.py
