+++ assemblerflow/generator/components/mapping.py
